[Refactor]: AnonymousPostProfileRepository 파일 이동
Related to: #559
